\name{R1W-package}
\alias{R1W-package}
\alias{R1W}
\docType{package}
\title{
\packageTitle{R1W}
}
\description{
\packageDescription{R1W}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{R1W}
\packageIndices{R1W}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{R1W}

Maintainer: \packageMaintainer{R1W}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
